feat(services): extract notification engine into @mulmoclaude/notifier
Fourth packages/services/* package — the active+history notification engine shared by
MulmoClaude and MulmoTerminal so both hosts run ONE engine (validation, write
coordinator, lifecycle rules, pub-sub + in-process fan-out) over their own workspaces
and pub-sub fabrics, instead of MulmoTerminal reimplementing it.

Injection seam (host-agnostic core, host specifics injected):
- configureNotifier({ writeJson, publishEvent, log? }) — IO deps. writeJson is the
  host's atomic JSON writer (kept single-sourced in the host with its Windows-retry
  rename); publishEvent binds the host's pub-sub channel; log adapts the host logger.
- setNotifierFilePaths({ active, history }) — file paths, split from the IO deps so the
  host binds production paths at module load while tests override them WITHOUT the deps
  call clobbering them. This preserves the existing test order
  (_setFilePathsForTesting → initNotifier), since initNotifier never touched paths.

Package layout: types.ts (dependency-free value types + constants), validate.ts (pure
publish-input validation, shared by engine + host HTTP route), store.ts (node:fs reads
+ injected atomic write), engine.ts (the coordinator + 10 methods + onEvent). Server-
only, dual ESM+CJS, no runtime deps. 9 package tests.

Host glue stays host-side (it binds the engine to host specifics): the plugin-runtime
API (runtime-api.ts), the HTTP route, the macOS reminder adapter, and the legacy
publishNotification wrapper — all unchanged.

- server/notifier/engine.ts reduced to a thin binding: re-exports the 10 methods +
  onEvent + validatePublishInput + NOTIFIER_LIMITS; initNotifier injects
  writeJsonAtomic + PUBSUB_CHANNELS.notifier + the host logger; _setFilePathsForTesting
  delegates to setNotifierFilePaths; production paths bound at module load.
- server/notifier/types.ts reduced to a re-export shim.
- server/notifier/store.ts deleted (orphaned — only the engine imported it).

Verification: typecheck:server/test/vue clean; the full original engine suite (66) plus
collections notifications (40), watcher (6), and optional-deps degradation (8) — 120
notifier-dependent tests — all green against the extracted package; package tests (9) green.
